add DTO

+++ lesson_05/users/users/src/main/java/de/ait/users/repository/UserRepository.java
@@ -7,7 +7,11 @@ import java.util.List;
 
 public interface UserRepository {
     List<User> findAll();
-    User findById(int id);
+
+    User findById(Long id);
+
     User save(User user);
+
     void delete(User user);
+
 }
